refactor(trusty-search): split orchestrator.rs into pollers/runner/finish to satisfy 500-SLOC cap
orchestrator.rs was 707 SLOC after the merge with main, violating the hard
production file cap. Split into four focused files:

  - orchestrator.rs  (~80 SLOC) — public entry points only (spawn_reindex,
                                   spawn_reindex_with_cleanup, collect_files)
  - pollers.rs       (~90 SLOC) — spawn_memory_poller + spawn_embedderd_rss_poller
  - runner.rs        (~270 SLOC) — Phase 1+2: walk, hash-cache load, pipelined
                                   parse/embed/commit batch loop
  - finish.rs        (~350 SLOC) — Phase 2 completion: prune pass, corpus swap,
                                   KG rebuild, terminal SSE event, GC cleanup

Public API surface unchanged; mod.rs re-exports are identical.
Ratchets .line-cap-allowlist.tsv: removes service/reindex/mod.rs entry (now 37 SLOC).
